skill system optimization.

+++ MuseumTycoon/Assets/Script/_Datas/ItemData.cs
@@ -28,7 +28,7 @@ public struct ItemData
         Name = _name;
         IAP_ID = _iAPId;
         if (_shoppingType == ShoppingType.RealMoney)
-            IAP_ID = Constant.instance.IAPIDCompany + Constant.instance.IAPIDGame + _itemType.ToString().ToLower() + "x" + _amount.ToString() + "_" + _shoppingType.ToString().ToLower() + "_" + _requiredMoney.ToString(); //com_kosippysudio_museumtycoon_gold5000x_realmoney_10
+            IAP_ID = Constant.instance.IAPIDCompany + Constant.instance.IAPIDGame + _itemType.ToString().ToLower() + "x" + _amount.ToString() + "_" + _shoppingType.ToString().ToLower() + "_" + ((int)_requiredMoney).ToString(); //com_kosippysudio_museumtycoon_gold5000x_realmoney_10
         Description = _description;
         RequiredMoney = _requiredMoney;
         Amount = _amount;
